# docs/swagger.yaml
basePath: /
definitions:
...
  system_instructions.CreateSystemInstructionRequest:
    properties:
      created_by:
        type: string
      is_active:
        type: boolean
      text:
        type: string
    type: object
  system_instructions.SystemInstruction:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      text:
        type: string
      version:
        type: integer
    type: object
  users.CreateUserRequest:
    properties:
      display_name:
        type: string
      email:
        type: string
      password:
        type: string
      role:
        $ref: '#/definitions/users.UserRoleType'
    type: object
  users.SignupRequest:
    properties:
      display_name:
        type: string
      email:
        type: string
      password:
        type: string
      role:
        $ref: '#/definitions/users.UserRoleType'
    type: object
  users.User:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
    type: object
  users.UserRoleType:
    enum:
    - ADMIN
    - INSTRUCTOR
    - LEARNER
    type: string
    x-enum-varnames:
    - UserRoleAdmin
    - UserRoleInstructor
    - UserRoleLearner
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API with password OAuth2 authentication
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms/
  title: Learning API
  version: "1.0"
paths:
...
  /signup:
    post:
      consumes:
      - application/json
      description: Create a new user account with email, password, and role
      parameters:
      - description: Signup Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.SignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/users.User'
        "400":
          description: invalid request
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Sign up a new user
      tags:
      - users
...
  /system-instructions:
    get:
      description: Get all system instructions
      responses:
        "200":
          description: List of system instructions
          schema:
            items:
              $ref: '#/definitions/system_instructions.SystemInstruction'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - read
      summary: List all system instructions
      tags:
      - System Instructions
    post:
      consumes:
      - application/json
      description: Create a new system instruction (immutable - cannot edit existing)
      parameters:
      - description: System instruction request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/system_instructions.CreateSystemInstructionRequest'
      responses:
        "201":
          description: Created system instruction
          schema:
            $ref: '#/definitions/system_instructions.SystemInstruction'
        "400":
          description: Bad request - invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - write
      summary: Create new system instruction
      tags:
      - System Instructions
  /system-instructions/{id}:
    get:
      description: Retrieve a specific system instruction by its UUID
      parameters:
      - description: Instruction ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: System instruction details
          schema:
            $ref: '#/definitions/system_instructions.SystemInstruction'
        "400":
          description: Bad request - invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Instruction not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - read
      summary: Get system instruction by ID
      tags:
      - System Instructions
  /system-instructions/{id}/activate:
    post:
      description: Mark a system instruction as active (deactivates other versions)
      parameters:
      - description: Instruction ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Activation status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request - invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - write
      summary: Activate system instruction
      tags:
      - System Instructions
  /system-instructions/active:
    get:
      description: Retrieve the currently active system instruction
      responses:
        "200":
          description: Active system instruction
          schema:
            $ref: '#/definitions/system_instructions.SystemInstruction'
        "404":
          description: Active instruction not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2:
        - read
      summary: Get active system instruction
      tags:
      - System Instructions
  /users:
    post:
      consumes:
      - application/json
      description: Create a new user with the provided details
      parameters:
      - description: Create User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/users.User'
        "400":
          description: invalid request
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      security:
      - OAuth2:
        - write
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a user from the system
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid request
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      security:
      - OAuth2:
        - write
      summary: Delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Retrieve a specific user by their unique ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.User'
        "400":
          description: invalid request
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      security:
      - OAuth2:
        - read
      summary: Get user by ID
      tags:
      - users
  /users/email/{email}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific user by their email address
      parameters:
      - description: User Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.User'
        "400":
          description: invalid request
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      security:
      - OAuth2:
        - read
      summary: Get user by email
      tags:
      - users
  /users/me:
    get:
      consumes:
      - application/json
      description: Retrieve the current user's profile from their JWT token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.User'
        "401":
          description: unauthorized
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      security:
      - OAuth2:
        - read
      summary: Get current authenticated user
      tags:
      - users
securityDefinitions:
  OAuth2:
    flow: password
    tokenUrl: /oauth/token
    type: oauth2
swagger: "2.0"